+++ seguranca-backend/src/main/java/br/gov/prefeitura/seguranca/service/UserDetailsImpl.java
@@ -16,7 +16,7 @@ import br.gov.prefeitura.seguranca.model.Usuario;
 public class UserDetailsImpl implements UserDetails {
 	private static final long serialVersionUID = 1L;
 
-	private Long id;
+	private Integer id;
 
 	private String username;
 
@@ -27,7 +27,7 @@ public class UserDetailsImpl implements UserDetails {
 
 	private Collection<? extends GrantedAuthority> authorities;
 
-	public UserDetailsImpl(Long id, String username, String email, String password,
+	public UserDetailsImpl(Integer id, String username, String email, String password,
 			Collection<? extends GrantedAuthority> authorities) {
 		this.id = id;
 		this.username = username;
@@ -54,7 +54,7 @@ public class UserDetailsImpl implements UserDetails {
 		return authorities;
 	}
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
